delete component moved and status label made to jsx file
