+++ shelfsense/src/util/services/ComponentService.jsx
@@ -4,32 +4,6 @@ import {destroyStoresAndLogout} from "../destroyStoresAndLogout.js";
 
 const BASE_URL = `${import.meta.env.VITE_API_URL}/components`;
 
-export const deleteComponent = async (id) => {
-    const BASE_URL = `${import.meta.env.VITE_API_URL}/components/${id}`;
-
-    try {
-        const response = await fetch(BASE_URL, {
-            method: "DELETE",
-            credentials: "include",
-        });
-
-        if (response.status === 401) {
-            await destroyStoresAndLogout();
-            return null;
-        }
-
-        if (!response.ok) {
-            console.log('Cannot delete component as it is associated with a product.');
-            return false;
-        }
-
-        return true;
-    } catch (error) {
-        console.error('Error occurred while deleting component:', error);
-        return null;
-    }
-};
-
 export const createApiRequest = async () => {
 
     const user = useSessionStore.getState().user;
